% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelskill-package.R
\docType{package}
\name{modelskill-package}
\alias{modelskill}
\alias{modelskill-package}
\title{Quantitative model skill indices and diagrams}
\description{
Compare numeric predictions against paired observations using
\code{\link[=model_metrics]{model_metrics()}} and \code{\link[=diagram_stats]{diagram_stats()}}. Visualize complementary aspects
of performance with \code{\link[=gg_taylor]{gg_taylor()}}, \code{\link[=gg_solar]{gg_solar()}} and \code{\link[=gg_target]{gg_target()}}.
All plots return ordinary ggplot objects. No spatial data class is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexandreWadoux/modelskill}
  \item Report bugs at \url{https://github.com/AlexandreWadoux/modelskill/issues}
}

}
\author{
\strong{Maintainer}: Alexandre M. J.-C. Wadoux \email{alexandre.wadoux@sydney.edu.au}

Authors:
\itemize{
  \item Alexandre M. J.-C. Wadoux \email{alexandre.wadoux@sydney.edu.au}
}

Other contributors:
\itemize{
  \item Modelskill contributors [contributor]
}

}
\keyword{internal}
