% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelskill-package.R
\docType{package}
\name{modelskill-package}
\alias{modelskill}
\alias{modelskill-package}
\title{Quantitative model skill indices and diagrams}
\description{
Compare numeric predictions against paired observations with
\code{\link[=model_metrics]{model_metrics()}} and individual metrics such as \code{\link[=rmse]{rmse()}} and \code{\link[=ccc]{ccc()}}.
Validate predictive uncertainty using \code{\link[=uncertainty_metrics]{uncertainty_metrics()}} and visualize
interval calibration with \code{\link[=gg_coverage]{gg_coverage()}}. Complementary graphical
diagnostics include \code{\link[=gg_taylor]{gg_taylor()}}, \code{\link[=gg_solar]{gg_solar()}} and \code{\link[=gg_target]{gg_target()}}. All plots
return ordinary ggplot objects. No spatial data class is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexandreWadoux/modelskill}
  \item Report bugs at \url{https://github.com/AlexandreWadoux/modelskill/issues}
}

}
\author{
\strong{Maintainer}: Alexandre M.J.-C. Wadoux \email{alexandre.wadoux@yahoo.fr} [copyright holder]

Authors:
\itemize{
  \item Alexandre M.J.-C. Wadoux \email{alexandre.wadoux@yahoo.fr} [copyright holder]
}

}
\keyword{internal}
